Добавил контроллеры для заказов и бронирования. Добавил сервис для отправки сообщений на почту.
