+++ server/package.json
@@ -2,18 +2,20 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "src/server.js",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "dev": "nodemon server.js"
+    "start": "node src/server.js",
+    "dev": "nodemon src/server.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "express": "^4.19.2",
+    "moment": "^2.30.1",
     "mongoose": "^8.2.1",
+    "nodemailer": "^6.9.14",
     "swagger-autogen": "^2.23.7"
   },
   "devDependencies": {
