--- server/server.js
@@ -1,33 +1,0 @@
-import express from 'express'
-import mongoose from 'mongoose'
-import { setupSwagger } from './swaggerConfig.js'
-import router from "./routes/index.js"
-
-const DB_URL = 'mongodb+srv://andrewrd0364:visagemain0364@cluster0.b5tzfiz.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0'
-
-const app = express()
-export const PORT = 5000
-
-app.use(express.json())
-app.use(express.static('images'))
-setupSwagger(app);
-
-app.use((req, res, next) => {
-    res.setHeader('Access-Control-Allow-Origin', '*');
-    res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE');
-    res.setHeader('Access-Control-Allow-Headers', 'Content-Type, Authorization');
-    next();
-});
-
-app.use('/api', router)
-
-async function startApp(){
-    try {
-        await mongoose.connect(DB_URL)
-        app.listen(PORT, () => console.log(`Server started on port ${PORT}`))
-    } catch (e) {
-        console.log(e)
-    }
-}
-
-startApp()
